# docs/api.yml
openapi: 3.0.3
info:
  title: Talevonia API
  version: "2.0.0"
  description: |
    Backend API for Talevonia MVP – a branching, AI-assisted personal storytelling app.

    **Authentication:**
    - Users authenticate via Apple Sign In (Google Sign In coming soon)
    - JWT-based authentication with short-lived access tokens (15 min) and long-lived refresh tokens (7 days)
    - All protected endpoints require Authorization header with Bearer token

    **User System:**
    - Users have accounts with user_id (UUID) as primary identifier
    - Account tiers: Free, Pro (determined by IAP subscriptions)
    - Welcome bonus: New users receive 5 credits (configurable) on first login with unique device

    **Credit System:**
    - Pro tier users receive 1000 subscription credits per month (expire monthly)
    - All users can purchase extra credits (100, 500, or 2000 packs) that never expire
    - Credits consumed in order: subscription first, then extra credits (FIFO by purchase date)
    - Credits replace daily quotas entirely
servers:
  - url: https://api.talevonia.app/api/v1
    description: Production
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login/apple or /auth/refresh
...
  schemas:
...
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    RefreshTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          example: 900
...
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
    ErrorObject:
      type: object
      properties:
        code:
          type: string
          example: QUOTA_EXCEEDED
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    # =============================================================================
    # Credits & IAP Schemas
    # =============================================================================
...